requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
pyTelegramBotAPI==4.13.0
python-dotenv
requests==2.31.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.4
